requirements_enhanced.txt
streamlit>=1.28.0
boto3>=1.34.0
PyPDF2>=3.0.0
python-docx>=0.8.11
pathlib
tempfile
datetime
json
os
typing